Fix unwanted behaviour because of command precedence.
The function 'reply' (that is used to reply to random messages or
to catch the netlist after /netlist is used) was used insted of
other commands because of his position in the code.
Now it is placed after all the others commands. Consequently, if any
existing command is used it is executed correctly insteod of triggering
the 'reply' function in advance.

 Changes to be committed:
	modified:   SpicePyBot.py